Force selectMainClass to just pick the first discovered one
This gets rid of asking the user on stdin, which hangs forever
on Windows. It might mess up a project that configures
selectMainClass itself, but few are likely to do that.

We do this in the probe rather than writing out a shim file
because we really don't want to change this if someone is
using regular sbt on a terminal.