Firewall IPC requests from ServerEngine
 * ServerEngine is now seen equally by ALL clients; requesting
   clients get the ID of the work item their request got associated
   with.
 * Requests may get associated with an existing work item, rather
   than creating a new one.
 * Requesters do not get the success/failure as a response, they
   have to watch events.
 * In implementing this, created a new intermediate queue for
   work items which is used to coalesce identical commands.

Interaction attempts (confirmation) now have to deal with the
possibility that multiple requesters asked for the same work.
Due to this problem this commit DOES NOT COMPILE (and ergo also
does not test).

Next commit will address interaction attempts.

One philosophy behind this commit is that the "response serial"
was abused/overloaded in old school sbt-remote-control. This serial
should from now on be used to match a SINGLE response to each request,
and the matched response should be one that happens in finite timeframe.
Events should not match to requests and we should never send multiple
responses.