Restore interaction prompts to working
Using replyTo here was not correct in the new order
of things. Wrapped the "work ID" Long in a WorkId case
class because otherwise this refactoring would not have
broken compilation if I missed a spot. This way we can't
confuse work IDs and serials.

Now we have to pick a client at random to ask for interaction;
if only one client has requested a given task, then it will
work the same way it did before, but if multiple have we
pick one at random.

It would probably be a little better to ask for confirmation from
all requesting clients, but probably not enough better to justify
the effort. In particular we'd have to have some kind of "cancel
interaction" request to the clients who didn't reply once one
of them did reply, so obsolete dialogs would get closed.