Move the work queue event stub so ServerState will be available
We want state.eventListeners to emit work queue changes, so
let's only emit the changes when we get the state because
we are about to execute. This will also have the effect
of compressing the change events if we want to.

We do not compress adds/removes, because we don't want to
execute some work without letting clients see it enqueued.