Use showFullKey when generating an ExecutionRequest from a key
This way we don't rely on the current project.

One potential downside of this is that we won't combine
duplicate requests in the same cases since we merge
requests using command string equality. But we probably need
to fix that more robustly anyway if it matters, which it
maybe won't really.