Fix task ID assignment in several ways
Use Task[_] not ScopedKey. ScopedKey had multiple problems:
 - multiple tasks with same key could be in an execution
 - some tasks don't have a key

Create an ID for the registered task only, in registered(),
not for its dependencies. Previous code was assuming registered
would be called only for the destination task not every task.

Add some sanity-checks (throw exceptions if we register a duplicate
ID or try to run a task with no ID).