Attempt to reply exactly once to all requests
This allows requesters to block for an "ack" or ErrorResponse if they see
fit. Without the non-error ack, there's no way to know whether an error
is going to arrive later. So this is good general practice to always
have a "yes I got it" reply even if for now we never send back an
error for a particular request.