Split apart watching build/values from requesting synthetic events
This allows the client to be more flexible and avoid extra events
when it doesn't need them. For example if it already has a local
cache it may not need to synthesize an event, etc.

One semantic change caused by this patch is that we now ALWAYS
synthesize a change event when we add any listener on the client
side, while before we only did it for the FIRST listener.