Fix CanKillServer integration test for withSbt changes
Because the connect count thing was changed to be number
of connects instead of number of connects OR errors,
the CanKillServer test was broken; reworked to test
in a different way.

Also, print expected stuff to stdout not stderr and
fix to have a start-small-and-increase timeout
instead of a start-large-and-decrease timeout.