When a new client connects, send the execution queue
Most of this patch is just rearranging TestExecution to
test it.

There's an important limitation that we don't currently
send the active execution, only the waiting ones.

The active one is going to be a little tricky, because
we need to guarantee that clients get either the
full waiting/started/finished event triple, or none
of it.

Minor thing in this patch is adding an assertion that
ServerEngineWork always has at least one requester,
because we use allRequesters.head unconditionally
when we dump the queue to a new client.