Log the .sbtserver/master.log filename in addition to connections/master.log
This required hoisting the filename up out of ServerEngine to SbtServer
so it could be shared. (Should the two "master.log" just be merged?
it would be easy to do building on this patch by sharing the logger
instead of just this filename. But for now just sharing the filename.)