Sync new clients to current execution state
This involves moving eventListeners to an eventSink object
because we need to synchronize adding listeners and sending
current state, to avoid sending events twice or missing events.

A small bugfix in here is that we did not send ExecutionWaiting
for synthetic execution requests previously. This bug was revealed
because the event state tracker throws an exception if events are
incorrect, so had to fix the bug to make the tests pass.