Attempt to make server exit or reboot cleanly
Both the queue from socket to ReadOnlyServerEngine, and the
queue from ReadOnlyServerEngine to ServerEngine, can now
send an "EOF" message. These tell the other one to exit.

The "exit" and "reload" sbt commands should now work I think;
so the SbtClient.requestSelfDestruct() API could be dropped, but
leaving it for now.

Various other shuffling to make it work.