Add BuildLoaded and BuildFailedToLoad events
For BuildLoaded you could also just watch for the first
BuildStructureChanged, but this is a little clearer if
what you really want is the first load.

From an SbtClient perspective, a reload and a restart of the
server JVM look the same, so for a given client connection
you would only ever get exactly one BuildLoaded or
BuildFailedToLoad in theory.