Make SbtClient require serializations in the constructor
Otherwise there is a race where the serializations might not
be added before the event we need to deserialize.

If we need more serializations on the fly, a possible API
would be to allow providing a deserializer to watch().
We could use that deserializer to rewrite events from
ValueChanged[JsValue] to ValueChanged[T] when invoking the
watch() handler. But not in this commit for now.