Rename ReadOnlyServerEngine to RequestProcessor
and nextStateRef to readOnlyStateRef, and clarify some comments.

These classes have evolved over time. The way it works now,
the ServerEngine is the sbt loop, and the request processor
class (now called RequestProcessor) happens to have a read
only copy of the state.

I think this will make it easier for newcomers to the code to
figure out what's going on.