Remove implicit genPickler genUnpickler from sbt.pickling._
This was in there by accident.

Now we have compilation failures in a bunch of new places where we need
to explicitly define missing picklers.

Had to rearrange CustomPicklersUnpicklers to make it compile
(implicits can't use implicits defined in a subtrait).
Had to copy-paste over a bunch of the primitive picklers.