Introduce "daemon mode" for clients, they don't keep server alive
This is intended to be for the Play fork plugin for example,
we don't want a dev-mode play server to keep sbt alive if
there aren't any IDEs or other user-visible clients open.
If Play wanted to be all clever it could set daemon mode
only after it goes idle itself (after no http requests for a
bit). Anyway, daemon mode can be toggled at any time so clients
can be as smart as they think is wise.