Add InspectRequest and ListSettingsRequest
This is supposed to allow displaying docs/tooltips on
tasks, and even displaying the entire settings graph, in
a client. Essentially an API for settings query.

InspectResponse is intended to have an optional element
that can be computed on the client side if the client
inspects all of the settings (loads the entire settings
graph). This is the "InspectPreanalysis"